fix: refactor confirmation directory
