+++ src/components/Confirmation/SignUpConfirmation/SignUpConfirmation.styles.js
