+++ src/components/Confirmation/SignUpConfirmation/index.js
@@ -5,7 +5,7 @@ import ReactMarkdown from 'react-markdown';
 
 import Text from 'components/Text';
 import { TEXT_TYPE } from 'components/Text/constants';
-import ConfirmationWrapper from 'components/Confirmation/ConfirmationWrapper';
+import ConfirmationWrapper from '../ConfirmationWrapper';
 
 import { styles } from './SignUpConfirmation.styles';
