+++ src/components/Confirmation/index.js
@@ -1,3 +1,5 @@
 export { default as ContactConfirmation } from './ContactConfirmation';
 export { default as SupplyConfirmation } from './SupplyConfirmation';
 export { default as ServiceConfirmation } from './ServiceConfirmation';
+export { default as SignUpConfirmation } from './SignUpConfirmation';
+export { default as OrganizationConfirmation } from './OrganizationConfirmation';
