+++ src/components/ServiceReview/Petcare.js
@@ -8,7 +8,7 @@ import { Routes } from 'constants/Routes';
 import { routeWithParams } from 'lib/utils/routes';
 
 import Text from 'components/Text';
-import SubRow from 'components/Confirmation/SubRow';
+import SubRow from 'components/SubRow';
 import { TEXT_TYPE } from 'components/Text/constants';
 
 import { styles } from './ServiceReview.styles';
