+++ src/containers/ServiceConfirmation.js
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
 
-import { ServiceConfirmation as ServiceConfirmationComponent } from 'components/Confirmation';
+import { ServiceConfirmation as ServiceConfirmationComponent } from 'components/Confirmation/ServiceConfirmation';
 
 function ServiceConfirmation({ service }) {
   return <ServiceConfirmationComponent service={service} />;
